OPENEUROPA-1659: Improvements on the Page theming + pull param on drone.
