+++ behat.yml.dist
@@ -51,6 +51,7 @@ default:
         page: "body"
         page header: ".ecl-page-header"
         page header site identity: ".ecl-page-header__identity"
+        page header intro: ".ecl-page-header__intro"
         priorities dropdown menu: "#nav-menu-expandable-group-3"
   formatters:
     progress: ~
