+++ modules/oe_theme_content_page/oe_theme_content_page.info.yml
@@ -1,0 +1,7 @@
+name: OpenEuropa Theme Page Content
+type: module
+description: Companion module for the OE Content Page module
+package: OpenEuropa
+core: 8.x
+dependencies:
+  - oe_content_page:oe_content_page
