+++ runner.yml.dist
@@ -14,14 +14,8 @@ drupal:
     profile: "minimal"
     name: "OpenEuropa"
   post_install:
-    - "./vendor/bin/drush en oe_theme_helper -y"
-    - "./vendor/bin/drush en oe_multilingual -y"
-    - "./vendor/bin/drush en oe_search -y"
-    - "./vendor/bin/drush en oe_corporate_blocks -y"
-    - "./vendor/bin/drush en oe_content -y"
-    - "./vendor/bin/drush en oe_content_page -y"
-    - "./vendor/bin/drush en config_devel -y"
-    - "./vendor/bin/drush en toolbar -y"
+    - "./vendor/bin/drush en oe_theme_helper oe_multilingual oe_search oe_corporate_blocks oe_content oe_content_page oe_theme_content_page -y"
+    - "./vendor/bin/drush en config_devel toolbar -y"
     - "./vendor/bin/drush theme:enable oe_theme -y"
     - "./vendor/bin/drush theme:enable seven -y"
     - "./vendor/bin/drush config-set system.theme default oe_theme -y"
