+++ tests/features/page-header.feature
@@ -12,7 +12,7 @@ Feature: Page header block component.
 
   Scenario Outline: The page header block shows the current page metadata.
     Given I am an anonymous user
-    When I go to the "<page>" demo page
+    When I go to the "<page>" page
     Then I should see the heading "<page>" in the "page header"
     And the breadcrumb trail should be "Home"
     And the breadcrumb active element should be "<page>"
@@ -32,7 +32,7 @@ Feature: Page header block component.
 
   Scenario: Change the title to verify that the page header is updated accordingly.
     Given I am logged in as a user that can "edit any" demo pages
-    When I go to the "Robots are everywhere" demo page
+    When I go to the "Robots are everywhere" page
     And I click "Edit"
     And I fill in "Title" with "Robots are everywhere nowadays"
     And I press "Save"
@@ -41,7 +41,15 @@ Feature: Page header block component.
     And the breadcrumb active element should be "Robots are everywhere nowadays"
 
     Given I am an anonymous user
-    When I go to the "Robots are everywhere nowadays" demo page
+    When I go to the "Robots are everywhere nowadays" page
     Then I should see the heading "Robots are everywhere nowadays" in the "page header"
     And the breadcrumb trail should be "Home"
     And the breadcrumb active element should be "Robots are everywhere nowadays"
+
+  Scenario: Page content type has custom metadata
+    Given "oe_page" content:
+      | title   | oe_page_summary | oe_content_content_owner                                                |
+      | My page | My summary      | http://publications.europa.eu/resource/authority/corporate-body/EP_PECH |
+    And I am an anonymous user
+    When I go to the "My page" page
+    Then I should see "My summary" in the "page header intro"
